- move MTE and manager to a more suitable package - fix the renderer - sprinkle in a couple ``@NotNull`s
