+++ src/main/java/gregtech/api/metatileentity/IWirelessCharger.java
@@ -3,11 +3,13 @@ package gregtech.api.metatileentity;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
 
+import org.jetbrains.annotations.NotNull;
+
 import java.util.List;
 
 public interface IWirelessCharger {
 
-    boolean canChargePlayerItems(EntityPlayer player);
+    boolean canChargePlayerItems(@NotNull EntityPlayer player);
 
-    void chargePlayerItems(List<ItemStack> stacksToCharge);
+    void chargePlayerItems(@NotNull List<ItemStack> stacksToCharge);
 }
