+++ src/main/java/gregtech/common/metatileentities/wirelesscharger/WirelessChargerManger.java
@@ -1,4 +1,4 @@
-package gregtech.api.metatileentity.registry;
+package gregtech.common.metatileentities.wirelesscharger;
 
 import gregtech.api.metatileentity.IWirelessCharger;
 import gregtech.api.util.GTUtility;
